Create week1&week2.py
Re: Fibonacci exercise responses
by DARREN YOUNG - Sunday, 28 January 2018, 8:26 PM
 
My name is Darren, so the first and last letter of my name (D + N = 4 + 14) give the number  18. The 18th Fibonacci number is 2584.

Re: Week 2 task
by DARREN YOUNG - Sunday, 4 February 2018, 10:43 PM
 
So in my example I ran with 3 scenarios:
1.    My surname is Young (Capital Y)
2.    My surname is young (all lowercase)
3.    My surname is YOUNG (All CAPS)
It can be seen from the program that ord () function will look at the first and last letter and return an integer/whole number from the Unicode code point.
My surname is Young
The first letter Y is number 89
The last letter g is number 103
Fibonacci number 192 is 5972304273877744135569338397692020533504

My surname is young
The first letter y is number 121
The last letter g is number 103
Fibonacci number 224 is 29090180355503362256910111038089984964854261893

My surname is YOUNG
The first letter Y is number 89
The last letter G is number 71
Fibonacci number 160 is 1226132595394188293000174702095995
So it can be seen that by changing the first and letter between lower case and uppercase, the ord function will return different numbers from the Unicode system. This in turn results in three different Fibonacci numbers.
What is the ord function?
Given a string of length one, return an integer representing the Unicode code point of the character when the argument is a unicode object, or the value of the byte when the argument is an 8-bit string. For example, ord('a') returns the integer 97, ord('€') (Euro sign) returns 8364.

What is Unicode? 
Unicode is a computing industry standard for the consistent encoding, representation, and handling of text expressed in most of the world's writing systems. The latest version contains a repertoire of 136,755 characters covering 139 modern and historic scripts, as well as multiple symbol sets.